+++ src/love/script.cpp
@@ -264,8 +264,6 @@ script::script(const std::string& file) {
 	chai.add(fun(&system::getOS), "getOS");
 	chai.add(fun(&system::getVersion), "getVersion");
 	chai.add(fun(&system::getVersionString), "getVersionString");
-	chai.add(fun<void, system, const std::string&, int>(&system::message), "message");
-	chai.add(fun<void, system, const std::string&>(&system::message), "message");
 
 	// Mouse
 	chai.add(fun(&mouse::getX), "getX");
@@ -286,6 +284,8 @@ script::script(const std::string& file) {
 	// Window
 	chai.add(fun(&window::setTitle), "setTitle");
 	chai.add(fun(&window::getTitle), "getTitle");
+	chai.add(fun<void, window, const std::string&, int>(&window::showMessageBox), "showMessageBox");
+	chai.add(fun<void, window, const std::string&>(&window::showMessageBox), "showMessageBox");
 
 	// Timer
 	chai.add(fun(&timer::getDelta), "getDelta");